:recycle: [Refactor] 패키지 명 변경 (대문자 -> 소문자)
