+++ src/main/java/kcs/funding/fundingboost/domain/repository/giftHubItem/GiftHubItemRepository.java
@@ -1,4 +1,4 @@
-package kcs.funding.fundingboost.domain.repository.GiftHubItem;
+package kcs.funding.fundingboost.domain.repository.giftHubItem;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface GiftHubItemRepository extends JpaRepository<GiftHubItem, Long>, GiftHubItemRepositoryCustom {
-    
+
     @Query("select g from GiftHubItem g" +
             " join fetch g.item i" +
             " join fetch g.member m" +
