+++ src/main/java/kcs/funding/fundingboost/domain/repository/giftHubItem/GiftHubItemRepositoryCustom.java
@@ -1,4 +1,4 @@
-package kcs.funding.fundingboost.domain.repository.GiftHubItem;
+package kcs.funding.fundingboost.domain.repository.giftHubItem;
 
 import java.util.List;
 import kcs.funding.fundingboost.domain.entity.GiftHubItem;
